All file are structured and .env file also added
