+++ src/app.js
@@ -7,9 +7,5 @@ app.use(express.json());
 
 connectToDB();
 
-app.use('/api/auth/student', require('./routes/student/studentLogin'));
-app.use('/api/auth/admin', require('./routes/admin/adminLogin'));
-
-
 app.use("/", userRoutes);
 module.exports = app;
