+++ src/routes/UserRoutes.js
@@ -4,5 +4,8 @@ const userController = require('../controllers/UserController');
 
 router.get('/', userController.home);
 
+router.post('/vyatha/api/studentLogin',userController.studentLogin);
+router.post('/vyatha/api/adminLogin',userController.adminLogin);
+
 
 module.exports = router;
